refactor(api): split admin routes into sub-module directory
Move monolithic admin.ts into admin/ directory structure to support
scaling to ~35 endpoints in Phase 3. No URL or behavior changes.

Closes #303
